refactor(store-core): extract session-list handlers into handlers/session-list.ts (audit P2-3)
Move the session_list family (handleSessionList, buildSessionListPatches,
cumulativeUsageEquals, chunkSubscribeSessionIds, SESSION_LIST_SUBSCRIBE_CHUNK_SIZE,
SessionListPatches) and the #4307 background_work_changed handler
(handleBackgroundWorkChanged, PendingBackgroundShellsBuilder) out of the
handlers/index.ts god-module into handlers/session-list.ts, behind the barrel's
existing export * re-export so the public surface is unchanged.

handleBackgroundWorkChanged moves with the session_list family because
buildSessionListPatches is its sole internal caller (it seeds each session's
pendingBackgroundShells from the snapshot). Drop the now-unused CumulativeUsage
and PendingBackgroundShell type imports from index.ts.

Pure move — bodies byte-identical to origin/main. index.ts 2929 -> 2573 lines.
